# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Trail Coach API
  description: >
    Personal AI trail-running coach for Denis Shuvalov.
    Receives Telegram messages, syncs Strava activities, generates coaching advice via Claude.
  version: 1.0.0
  contact:
    name: TaurDT
servers:
  - url: https://your-app.azurecontainerapps.io
    description: Production (Azure Container Apps)
  - url: http://localhost:3000
    description: Local development

# ---------------------------------------------------------------------------
# Security schemes
# ---------------------------------------------------------------------------
components:
  securitySchemes:
    TelegramWebhookSecret:
      type: apiKey
      in: header
      name: X-Telegram-Bot-Api-Secret-Token
      description: Secret token set when registering the Telegram webhook.

    StravaWebhookSecret:
      type: apiKey
      in: path
      name: secret
      description: >
        Secret path segment (`COMPOSIO_WEBHOOK_SECRET`) embedded in the URL.
        Strava and setup endpoints use this for authentication.

  # ---------------------------------------------------------------------------
  # Reusable schemas
  # ---------------------------------------------------------------------------
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
          example: Unauthorized
    OkResponse:
      type: object
      required: [ok]
      properties:
        ok:
          type: boolean
          example: true

    # --- Telegram ---
    TelegramUser:
      type: object
      properties:
        id:
          type: integer
        is_bot:
          type: boolean
        first_name:
          type: string
        username:
          type: string
        language_code:
          type: string
    TelegramChat:
      type: object
      required: [id, type]
      properties:
        id:
          type: integer
          description: Unique identifier for this chat.
        type:
          type: string
          enum: [private, group, supergroup, channel]
        first_name:
          type: string
        username:
          type: string
    TelegramMessage:
      type: object
      required: [message_id, chat, date]
      properties:
        message_id:
          type: integer
        from:
          $ref: "#/components/schemas/TelegramUser"
        chat:
          $ref: "#/components/schemas/TelegramChat"
        date:
          type: integer
          description: Unix timestamp when the message was sent.
        text:
          type: string
          description: Text content of the message. Only text messages are processed; others are silently ignored.
          example: "Как прошла моя тренировка вчера?"
    TelegramUpdate:
      type: object
      required: [update_id]
      properties:
        update_id:
          type: integer
          description: Unique update identifier.
        message:
          $ref: "#/components/schemas/TelegramMessage"

    # --- Strava webhook ---
...
paths:
...
  /webhook/telegram:
    post:
      summary: Telegram message delivery
      description: >
        Telegram sends all bot updates here. The server responds **200 immediately** to
        prevent Telegram retries, then processes the message asynchronously:
        syncs Strava activities, queries Supabase for the last 30 days, calls Claude,
        and sends the coaching reply back via Telegram Bot API.

        Only `message.text` updates are processed; all others are silently ignored.
      operationId: postTelegramWebhook
      tags: [Telegram]
      security:
        - TelegramWebhookSecret: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TelegramUpdate"
            example:
              update_id: 100500
              message:
                message_id: 42
                from:
                  id: 123456789
                  is_bot: false
                  first_name: Denis
                  username: taurdt
                  language_code: ru
                chat:
                  id: 123456789
                  type: private
                  first_name: Denis
                  username: taurdt
                date: 1746270000
                text: "Как прошла моя тренировка вчера?"
      responses:
        "200":
          description: >
            Always returned immediately upon receiving a valid request.
            Actual coaching reply is sent asynchronously via Telegram Bot API.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
        "401":
          description: Missing or invalid `X-Telegram-Bot-Api-Secret-Token` header.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # --- Strava webhook ---
...